+++ .travis.yml
@@ -17,4 +17,14 @@ after_success:
 env:
   global:
   - SAUCE_USERNAME: Desire2Learn
-<%= deployInfo %>
+  - REPO_NAME=user-profile-card
+  - OWNER_NAME=BrightspaceUILabs
+deploy:
+  - provider: npm
+    email: d2ltravisdeploy@d2l.com
+    skip_cleanup: true
+    api_key:
+      # d2l-travis-deploy: ...
+    on:
+      tags: true
+      repo: BrightspaceUILabs/user-profile-card
