+++ LICENSE
@@ -186,7 +186,7 @@ APPENDIX: How to apply the Apache License to your work.
 	same "printed page" as the copyright notice for easier
 	identification within third-party archives.
 
-Copyright <%= year %> D2L Corporation
+Copyright 2020 D2L Corporation
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
