+++ README.md
@@ -1,58 +1,117 @@
-# lit-element-template
+# d2l-labs-user-profile-card
 
-Template for creating BrightspaceUI lit elements.
+[![NPM version](https://img.shields.io/npm/v/@brightspace-ui-labs/user-profile-card.svg)](https://www.npmjs.org/package/@brightspace-ui-labs/user-profile-card)
+[![Dependabot badge](https://flat.badgen.net/dependabot/BrightspaceUILabs/user-profile-card?icon=dependabot)](https://app.dependabot.com/)
+[![Build status](https://travis-ci.com/@brightspace-ui-labs/user-profile-card.svg?branch=master)](https://travis-ci.com/@brightspace-ui-labs/user-profile-card)
 
-With this template you get:
+> Note: this is a ["labs" component](https://github.com/BrightspaceUI/guide/wiki/Component-Tiers). While functional, these tasks are prerequisites to promotion to BrightspaceUI "official" status:
+>
+> - [ ] [Design organization buy-in](https://github.com/BrightspaceUI/guide/wiki/Before-you-build#working-with-design)
+> - [ ] [design.d2l entry](http://design.d2l/)
+> - [ ] [Architectural sign-off](https://github.com/BrightspaceUI/guide/wiki/Before-you-build#web-component-architecture)
+> - [ ] [Continuous integration](https://github.com/BrightspaceUI/guide/wiki/Testing#testing-continuously-with-travis-ci)
+> - [ ] [Cross-browser testing](https://github.com/BrightspaceUI/guide/wiki/Testing#cross-browser-testing-with-sauce-labs)
+> - [ ] [Unit tests](https://github.com/BrightspaceUI/guide/wiki/Testing#testing-with-polymer-test) (if applicable)
+> - [ ] [Accessibility tests](https://github.com/BrightspaceUI/guide/wiki/Testing#automated-accessibility-testing-with-axe)
+> - [ ] [Visual diff tests](https://github.com/BrightspaceUI/visual-diff)
+> - [ ] [Localization](https://github.com/BrightspaceUI/guide/wiki/Localization) with Serge (if applicable)
+> - [ ] Demo page
+> - [ ] README documentation
 
-* Project boilerplate including: README, .editorconfig, .gitignore, package.json, CODEOWNERS and LICENSE (Apache-2.0)
-* A basic LitElement scaffold
-* Demo page for the element
-* Test page for the element
-* Travis CI ready-to-go
-* Local tests that do linting using ESLint and unit tests
-* Cross-browser testing from Travis CI using Sauce Labs
-* npm publish setup
+day light user profile card
 
-## Setup
+## Installation
 
-This assumes you have node installed.
+To install from NPM:
 
-1. Follow the GitHub instructions [here](https://help.github.com/en/articles/creating-a-repository-from-a-template) to create a new repository from this template, then clone the new repository on your local machine.
-2. Run the script, entering information as prompted (e.g., for repo name, description, etc.):
+```shell
+npm install @brightspace-ui-labs/user-profile-card
 ```
-node setup/configure-repo.js
+
+## Usage
+
+```html
+<script type="module">
+    import '@brightspace-ui-labs/user-profile-card/user-profile-card.js';
+</script>
+<d2l-labs-user-profile-card>My element</d2l-labs-user-profile-card>
 ```
 
-After the script successfully runs, follow the instructions on the generated README for usage of your new component. The setup directory can now be removed from your component repo.
+**Properties:**
 
-### Sauce Labs
+| Property | Type | Description |
+|--|--|--|
+| | | |
 
-To do cross-browser testing using Sauce Labs, the API key needs to be encrypted into the .travis.yml file.
+**Accessibility:**
 
-To learn more about how to set this up, see the [testing](https://github.com/BrightspaceUI/guide/wiki/Testing) section of The Guide.
+To make your usage of `d2l-labs-user-profile-card` accessible, use the following properties when applicable:
+
+| Attribute | Description |
+|--|--|
+| | |
+
+## Developing, Testing and Contributing
+
+After cloning the repo, run `npm install` to install dependencies.
+
+### Running the demos
+
+To start an [es-dev-server](https://open-wc.org/developing/es-dev-server.html) that hosts the demo page and tests:
+
+```shell
+npm start
+```
+
+### Linting
+
+```shell
+# eslint and lit-analyzer
+npm run lint
+
+# eslint only
+npm run lint:eslint
+
+# lit-analyzer only
+npm run lint:lit
+```
+
+### Testing
+
+```shell
+# lint, unit test and visual-diff test
+npm test
+
+# lint only
+npm run lint
+
+# unit tests only
+npm run test:headless
+
+# debug or run a subset of local unit tests
+# then navigate to `http://localhost:9876/debug.html`
+npm run test:headless:watch
+```
 
 ### Visual Diff Testing
 
-To do visual diff testing, the `GITHUB_RELEASE_TOKEN`, `VISUAL_DIFF_S3_ID`, and `VISUAL_DIFF_S3_SECRET` need to be encrypted into the .travis.yml file.
+This repo uses the [@brightspace-ui/visual-diff utility](https://github.com/BrightspaceUI/visual-diff/) to compare current snapshots against a set of golden snapshots stored in source control.
 
-To learn more about how to set this up, see the [Running in CI](https://github.com/BrightspaceUI/visual-diff#running-in-ci) section of the visual-diff readme.
+```shell
+# run visual-diff tests
+npm run test:diff
 
-To use the Visual Difference Bot to run visual diff testing, see [here](https://github.com/BrightspaceUI/visual-difference-bot/blob/master/README.md/#utilizing-the-deployed-bot-for-your-existing-brightspace-repo).
+# subset of visual-diff tests:
+npm run test:diff -- -g some-pattern
 
-### Versioning
+# update visual-diff goldens
+npm run test:diff:golden
+```
 
-A `GITHUB_RELEASE_TOKEN` needs to be encrypted into the .travis.yml file in order for the auto-versioning mentioned in the `Versioning, Releasing & Deploying` section of the generated README to work.
+Golden snapshots in source control must be updated by Travis CI. To trigger an update, press the "Regenerate Goldens" button in the pull request `visual-difference` test run.
 
-### Publishing
+## Versioning, Releasing & Deploying
 
-If you use `'publish': true` in your `config` object in configure-repo.js, your package.json and .travis.yml files will be setup for future public publishing with `npm`. To complete getting the publishing to work:
-* A `d2l-travis-deploy` API key needs to be encrypted and added into the .travis.yml file under `# d2l-travis-deploy: ...` (replace the `...` with the token used for generating the key).
-* Once you are ready to publish your element, navigate to the root directory of your package and use the command `npm publish --access public` (see [here](https://docs.npmjs.com/creating-and-publishing-scoped-public-packages#publishing-scoped-public-packages) for more information).
+All version changes should obey [semantic versioning](https://semver.org/) rules.
 
-### Localization
-
-To have localization setup by the configure-repo.js script, type "yes" to the Localization prompt, then select if resources should be statically or dynamically imported. See [here](https://github.com/BrightspaceUI/core/blob/master/mixins/localize-mixin.md#language-resources) for more information.
-
-## Developing and Contributing
-
-Pull requests welcome!
+Include either `[increment major]`, `[increment minor]` or `[increment patch]` in your merge commit message to automatically increment the `package.json` version, create a tag, and trigger a deployment to NPM.
