+++ test/user-profile-card.visual-diff.js
@@ -1,9 +1,9 @@
 const puppeteer = require('puppeteer');
 const VisualDiff = require('@brightspace-ui/visual-diff');
 
-describe('<%= name %>', () => {
+describe('d2l-labs-user-profile-card', () => {
 
-	const visualDiff = new VisualDiff('<%= shortName %>', __dirname);
+	const visualDiff = new VisualDiff('user-profile-card', __dirname);
 
 	let browser, page;
 
@@ -11,7 +11,7 @@ describe('<%= name %>', () => {
 		browser = await puppeteer.launch();
 		page = await visualDiff.createPage(browser);
 		await page.setViewport({width: 800, height: 800, deviceScaleFactor: 2});
-		await page.goto(`${visualDiff.getBaseUrl()}/test/<%= shortName %>.visual-diff.html`, {waitUntil: ['networkidle0', 'load']});
+		await page.goto(`${visualDiff.getBaseUrl()}/test/user-profile-card.visual-diff.html`, {waitUntil: ['networkidle0', 'load']});
 		await page.bringToFront();
 	});
